Fix behavior when no start directory given
According to the readme if no start directory is given
then it will start searching from the current directory.
This was not the case since it started searching from the
parent directory where the `require-dot-file` resides. This
breaks in `semantic-ui` when `require-dot-file` is installed
non locally either globally or via the `nix` package manager.